+++ bcbio/ngsalign/star.py
@@ -58,6 +58,7 @@ def align(fastq_file, pair_file, ref_file, names, align_dir, data):
             "--outStd BAM_Unsorted {srna_opts} "
             "--limitOutSJcollapsed 2000000 "
             "--outSAMtype BAM Unsorted "
+            "--outSAMmapqUnique 60 "
             "--outSAMunmapped Within --outSAMattributes %s " % " ".join(ALIGN_TAGS))
         cmd += _add_sj_index_commands(fastq_file, ref_file, gtf_file) if not srna else ""
         cmd += " --readFilesCommand zcat " if is_gzipped(fastq_file) else ""